refactor: move providers/errors.rs into a module inside providers/mod.rs
